fix: rename components folder to remove hidden Cyrillic character
