+++ src/modules/components/ProductDetailsPage/components/ProductCardDetails/components/ProductAbout/ProductAbout.module.scss
