+++ src/modules/components/ProductDetailsPage/components/ProductCardDetails/components/ProductAbout/ProductAbout.tsx
