+++ src/modules/components/ProductDetailsPage/components/ProductCardDetails/components/ProductAbout/ProductAboutSkeleton.tsx
