+++ src/modules/components/ProductDetailsPage/components/ProductCardDetails/components/ProductAbout/index.ts
