+++ src/modules/components/ProductDetailsPage/components/ProductCardDetails/components/ProductActions/ProductActions.module.scss
