+++ src/modules/components/ProductDetailsPage/components/ProductCardDetails/components/ProductGallery/ProductGallery.module.scss
