+++ src/modules/components/ProductDetailsPage/components/ProductCardDetails/components/ProductGallery/ProductGallerySkeleton.tsx
