+++ src/modules/components/ProductDetailsPage/components/ProductCardDetails/components/ProductGallery/index.ts
