+++ src/modules/components/ProductDetailsPage/components/ProductCardDetails/components/ProductTechSpecs/ProductTechSpecs.module.scss
