+++ src/modules/components/ProductDetailsPage/components/ProductCardDetails/components/ProductTechSpecs/ProductTechSpecsSkeleton.tsx
