+++ src/modules/components/ProductDetailsPage/components/ProductCardDetails/components/ProductTechSpecs/index.ts
